Implement shared policy wizard
